v0.10.0: tech-debt sweep — import-bug fix, CI, ruff, cli.py split
Fixed:
- Latent clean-install crash: millet/{capture,audio,utils,languages}.py
  shims imported from meet_record (pre-rename pkg). A clean
  'pip install millet-pipeline' (depends on millet-record -> millet_record)
  raised ModuleNotFoundError: meet_record. Now import from millet_record.
  voiceprint.py lazy imports switched meet.* -> millet.* (+ dropped 2
  unused). New tests/test_shim_imports.py guards regression.

Changed:
- cli.py (1929 lines) split into a cli/ package: one module per command
  + cli/_helpers.py; cli/__init__.py defines the main group and
  re-exports every command symbol so millet.subcommands /
  meet.subcommands entry points keep resolving.
- CI fixed: linted/tested dead meet/ paths (pkg is millet/) and installed
  meetscribe-record (old name). Now lints millet/+tests/, installs
  millet-record, runs the no-torch/no-GTK suites.
- Added [tool.ruff] config (mirrors vezir's ruleset); cleaned ~120
  findings. Rewrote legacy meet.* test refs -> millet.* (recovered ~29
  previously-erroring tests).

No user-facing behavior change; minor bump for the cli/ package split.
Tests: 233 passed, 7 env-only fails (GTK/torch); ruff clean.

+++ millet/frontmatter.py
@@ -226,7 +226,7 @@ def _iso_duration(seconds: float | None) -> str | None:
     """
     if seconds is None or seconds <= 0:
         return None
-    total = int(round(seconds))
+    total = round(seconds)
     h, rem = divmod(total, 3600)
     m, s = divmod(rem, 60)
     parts = []
